requirements.txt
# Core
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.13.0
matplotlib>=3.9.0
seaborn>=0.13.0
tqdm>=4.66.0
pyyaml>=6.0.1

# ECG Processing
wfdb>=4.1.0
neurokit2>=0.2.9

# PyTorch (install separately with CUDA if you have a GPU)
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0

# Quantization & Export
onnx>=1.16.0
onnxruntime>=1.17.0

# Interpretability
captum>=0.7.0
shap>=0.45.0
lime>=0.2.0.1

# ML Utilities
scikit-learn>=1.5.0
imbalanced-learn>=0.12.0

# Notebook support
jupyter>=1.0.0
ipykernel>=6.29.0

# Dashboard
streamlit>=1.37.0
